Add discount handling to the Proposal class
A new Discounts class was introduced to keep track of various discounted plans and the amount of discount. The Proposal class has been updated to handle these discounts - it can determine if an InsuredGroup is eligible for a discount, calculate the total cost considering discounts and handle adding and removing discounts. Relevant